+++ TinyPG/Examples/TestCpp/SimpleExpr/src/Scanner.cpp
@@ -142,7 +142,7 @@ namespace TinyPG
 
 			tok = Token(startpos, endpos);
 
-			for (i = 0; i < scantokens.size(); i++)
+			for (i = 0; i < (int)scantokens.size(); i++)
 			{
 				std::regex r = Patterns[scantokens[i]];
 				std::smatch m;
@@ -164,7 +164,7 @@ namespace TinyPG
 			}
 			else if (tok.StartPos == tok.EndPos)
 			{
-				if (tok.StartPos < Input.length())
+				if (tok.StartPos < (int)Input.length())
 					tok.Text = Input.substr(tok.StartPos, 1);
 				else
 					tok.Text = "EOF";
@@ -172,7 +172,7 @@ namespace TinyPG
 
 			// Update the line and column count for error reporting.
 			tok.Line = currentline;
-			if (tok.StartPos < Input.length())
+			if (tok.StartPos < (int)Input.length())
 				tok.Column = tok.StartPos - (int)Input.find_last_of('\n', tok.StartPos);
 
 			if (contains(SkipList, tok.Type))
