+++ TinyPG/Examples/TestCpp/TinyExe/src/Scanner.cpp
@@ -258,7 +258,7 @@ namespace TinyExe
 
 			tok = Token(startpos, endpos);
 
-			for (i = 0; i < scantokens.size(); i++)
+			for (i = 0; i < (int)scantokens.size(); i++)
 			{
 				std::regex r = Patterns[scantokens[i]];
 				std::smatch m;
@@ -280,7 +280,7 @@ namespace TinyExe
 			}
 			else if (tok.StartPos == tok.EndPos)
 			{
-				if (tok.StartPos < Input.length())
+				if (tok.StartPos < (int)Input.length())
 					tok.Text = Input.substr(tok.StartPos, 1);
 				else
 					tok.Text = "EOF";
@@ -288,7 +288,7 @@ namespace TinyExe
 
 			// Update the line and column count for error reporting.
 			tok.Line = currentline;
-			if (tok.StartPos < Input.length())
+			if (tok.StartPos < (int)Input.length())
 				tok.Column = tok.StartPos - (int)Input.find_last_of('\n', tok.StartPos);
 
 			if (contains(SkipList, tok.Type))
